# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ABC",
        "description": "Methods of sale endpoint",
        "contact": {
            "name": "José G Romano",
            "email": "jose_guilherme_romano@hotmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sale/{id}": {
            "get": {
                "summary": "Get a specific sale",
                "operationId": "33068aacbfdcf9274795e31578d8faa2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sale",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a single sale",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sale_id": {
                                            "type": "string",
                                            "example": "202403051"
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": "35000"
                                        },
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Product 1"
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "example": "3.500"
                                                    },
                                                    "amount": {
                                                        "type": "integer",
                                                        "example": "10"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No sale found"
                    }
                }
            },
            "put": {
                "summary": "Update a sale",
                "operationId": "2a65d0740cb94e6a7c0ee8c778ad6c95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sale",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "products"
                                ],
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "amount"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "amount": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sale updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Product 1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No sale found"
                    }
                }
            }
        },
        "/api/sale": {
            "post": {
                "summary": "Create a new sale",
                "operationId": "a4679729ade7c8239f4311330b84da7f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "products"
                                ],
                                "properties": {
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "amount"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "amount": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Product 1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sale/{id}": {
            "delete": {
                "summary": "Cancel a sale by ID (using SoftDeletes)",
                "description": "Deletes a sale from the database by its ID.",
                "operationId": "deleteSaleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sale to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sale deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Product 1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sale not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No Sale found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}